Refactor server_status_processes to static methods
