+++ server_status_processes.php
@@ -8,10 +8,10 @@
 
 use PhpMyAdmin\Response;
 use PhpMyAdmin\Server\Status\Data;
+use PhpMyAdmin\Server\Status\Processes;
 
 require_once 'libraries/common.inc.php';
 require_once 'libraries/server_common.inc.php';
-require_once 'libraries/server_status_processes.lib.php';
 
 /**
  * Replication library
@@ -47,7 +47,7 @@ if ($response->isAjax() && !empty($_REQUEST['kill'])) {
     $response->addJSON('message', $message);
 } elseif ($response->isAjax() && !empty($_REQUEST['refresh'])) {
     // Only sends the process list table
-    $response->addHTML(PMA_getHtmlForServerProcessList());
+    $response->addHTML(Processes::getHtmlForServerProcesslist());
 } else {
     // Load the full page
     $header   = $response->getHeader();
@@ -55,9 +55,9 @@ if ($response->isAjax() && !empty($_REQUEST['kill'])) {
     $scripts->addFile('server_status_processes.js');
     $response->addHTML('<div>');
     $response->addHTML($serverStatusData->getMenuHtml());
-    $response->addHTML(PMA_getHtmlForProcessListFilter());
-    $response->addHTML(PMA_getHtmlForServerProcesslist());
-    $response->addHTML(PMA_getHtmlForProcessListAutoRefresh());
+    $response->addHTML(Processes::getHtmlForProcessListFilter());
+    $response->addHTML(Processes::getHtmlForServerProcesslist());
+    $response->addHTML(Processes::getHtmlForProcessListAutoRefresh());
     $response->addHTML('</div>');
 }
 exit;
